+++ tests/src/simbio/se/sau/test/TestSauLibrary.java
@@ -10,6 +10,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
 
+import simbio.se.sau.encryption.Encryption;
 import simbio.se.sau.json.JsonUtils;
 import simbio.se.sau.log.SimbiLog;
 import simbio.se.sau.persistense.PreferencesHelper;
@@ -242,6 +243,17 @@ public class TestSauLibrary extends AndroidTestCase {
 		keyKeyMap.put(null, key3);
 		assertEquals(keyKeyMap.get(null), key3);
 		assertNull(keyKeyMap.get(key3));
+
+		// assert encrypt
+		Encryption encryption = new Encryption();
+		String key = "Some Key";
+		String data = "top secret string";
+
+		String encrypted = encryption.encrypt(key, data);
+		SimbiLog.print(encrypted);
+
+		String decrypted = encryption.decrypt(key, encrypted);
+		SimbiLog.print(decrypted);
 	}
 
 	/**